# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ent Schema API",
    "description": "This is an auto generated API description made out of an Ent schema definition",
    "version": "0.1.0"
  },
  "paths": {
    "/admin-areas": {
      "get": {
        "tags": [
          "AdminArea"
        ],
        "summary": "List AdminAreas",
        "description": "List AdminAreas.",
        "operationId": "listAdminArea",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer",
              "maximum": 255,
              "minimum": 1
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Name of the administrative area",
            "schema": {
              "type": "string",
              "maxLength": 255,
              "minLength": 2
            }
          },
          {
            "name": "abbr",
            "in": "query",
            "description": "Abbreviation of the administrative area, can be a CSV list",
            "schema": {
              "type": "string",
              "maxLength": 255,
              "minLength": 2
            }
          },
          {
            "name": "trashed",
            "in": "query",
            "description": "Whether to include trashed items",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of administrative areas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "current_page": {
                      "description": "Page number (1-based)",
                      "type": "integer",
                      "minimum": 1
                    },
                    "total": {
                      "description": "Total number of items",
                      "type": "integer",
                      "minimum": 0
                    },
                    "per_page": {
                      "description": "Number of items per page",
                      "type": "integer",
                      "minimum": 1
                    },
                    "last_page": {
                      "description": "Last page number",
                      "type": "integer",
                      "minimum": 1
                    },
                    "from": {
                      "description": "Index (1-based) of the first item in the current page",
                      "type": "integer",
                      "minimum": 0
                    },
                    "to": {
                      "description": "Index (1-based) of the last item in the current page",
                      "type": "integer",
                      "minimum": 0
                    },
                    "first_page_url": {
                      "description": "URL to the first page",
                      "type": "string"
                    },
                    "last_page_url": {
                      "description": "URL to the last page",
                      "type": "string"
                    },
                    "next_page_url": {
                      "description": "URL to the next page",
                      "type": "string"
                    },
                    "prev_page_url": {
                      "description": "URL to the previous page",
                      "type": "string"
                    },
                    "path": {
                      "description": "Base path of the request",
                      "type": "string"
                    },
                    "data": {
                      "description": "List of administrative areas",
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminAreaList"
                      }
                    }
                  },
                  "required": [
                    "current_page",
                    "total",
                    "per_page",
                    "last_page",
                    "from",
                    "to",
                    "first_page_url",
                    "last_page_url",
                    "next_page_url",
                    "prev_page_url",
                    "path",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "AdminArea"
        ],
        "summary": "Create a new AdminArea",
        "description": "Creates a new AdminArea and persists it to storage.",
        "operationId": "createAdminArea",
        "requestBody": {
          "description": "AdminArea to create",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "parent_id": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "name": {
                    "description": "Administrative area name",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                  },
                  "abbr": {
                    "description": "Administrative area abbreviations, CSV values",
                    "type": "string",
                    "nullable": true,
                    "maxLength": 255,
                    "minLength": 1
                  }
                },
                "additionalProperties": false,
                "required": [
                  "name"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "AdminArea created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAreaCreate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminAreaCreate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "parent_id": {
            "type": "integer",
            "minimum": 1
          },
          "name": {
            "description": "Administrative area name",
            "type": "string",
            "maxLength": 255,
            "minLength": 2
          },
          "abbr": {
            "description": "Administrative area abbreviations, CSV values",
            "type": "string",
            "nullable": true,
            "maxLength": 255,
            "minLength": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "AdminAreaList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "parent_id": {
            "type": "integer",
            "minimum": 1
          },
          "name": {
            "description": "Administrative area name",
            "type": "string",
            "maxLength": 255,
            "minLength": 2
          },
          "abbr": {
            "description": "Administrative area abbreviations, CSV values",
            "type": "string",
            "nullable": true,
            "maxLength": 255,
            "minLength": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "invalid input, data invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
...
      "404": {
        "description": "resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
      "409": {
        "description": "conflicting resources",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
      "500": {
        "description": "unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      }
    }
  }
}